refactor(providers): centralize metadata and config parsing
