+++ src/llm/provider-profile.ts
@@ -1,142 +1,20 @@
-import type { CliProvider } from "../config.js";
-import {
-  buildGitHubModelsHeaders,
-  GITHUB_MODELS_BASE_URL,
-  resolveGitHubModelsApiKey,
-} from "./github-models.js";
+import { buildGitHubModelsHeaders, resolveGitHubModelsApiKey } from "./github-models.js";
 import { normalizeGatewayStyleModelId, parseGatewayStyleModelId } from "./model-id.js";
 import type { ModelRequestOptions } from "./model-options.js";
 import { resolveOpenAiClientConfig } from "./openai-client-config.js";
+import {
+  CLI_PROVIDER_PROFILES,
+  GATEWAY_PROVIDER_PROFILES,
+  getCliProviderProfile,
+  getGatewayProviderProfile,
+  parseCliProviderName,
+  type CliProvider,
+  type CliRequiredModelEnv,
+  type GatewayProvider,
+  type GatewayRequiredModelEnv,
+  type RequiredModelEnv,
+} from "./provider-registry.js";
 import type { OpenAiClientConfig } from "./providers/types.js";
-
-export type GatewayProvider =
-  | "xai"
-  | "openai"
-  | "google"
-  | "anthropic"
-  | "zai"
-  | "nvidia"
-  | "minimax"
-  | "github-copilot"
-  | "ollama";
-
-export type RequiredModelEnv =
-  | "XAI_API_KEY"
-  | "OPENAI_API_KEY"
-  | "NVIDIA_API_KEY"
-  | "GEMINI_API_KEY"
-  | "ANTHROPIC_API_KEY"
-  | "OPENROUTER_API_KEY"
-  | "Z_AI_API_KEY"
-  | "MINIMAX_API_KEY"
-  | "GITHUB_TOKEN"
-  | "OLLAMA_BASE_URL"
-  | "CLI_CLAUDE"
-  | "CLI_CODEX"
-  | "CLI_GEMINI"
-  | "CLI_AGENT"
-  | "CLI_OPENCLAW"
-  | "CLI_OPENCODE"
-  | "CLI_COPILOT"
-  | "CLI_AGY"
-  | "CLI_PI";
-
-export type ProviderExecution =
-  | "simple"
-  | "google"
-  | "anthropic"
-  | "openai-http"
-  | "openai-compatible";
-
-export type GatewayProviderProfile = {
-  requiredEnv: RequiredModelEnv;
-  execution: ProviderExecution;
-  supportsDocuments: boolean;
-  supportsStreaming: boolean;
-  supportsVideoUnderstanding: boolean;
-  defaultBaseUrl?: string;
-  forceChatCompletions?: boolean;
-};
-
-export const DEFAULT_OLLAMA_BASE_URL = "http://localhost:11434/v1";
-
-export const DEFAULT_MINIMAX_BASE_URL = "https://api.minimax.io/v1";
-
-const GATEWAY_PROVIDER_PROFILES: Record<GatewayProvider, GatewayProviderProfile> = {
-  xai: {
-    requiredEnv: "XAI_API_KEY",
-    execution: "simple",
-    supportsDocuments: false,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-  },
-  openai: {
-    requiredEnv: "OPENAI_API_KEY",
-    execution: "openai-http",
-    supportsDocuments: true,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-  },
-  google: {
-    requiredEnv: "GEMINI_API_KEY",
-    execution: "google",
-    supportsDocuments: true,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: true,
-  },
-  anthropic: {
-    requiredEnv: "ANTHROPIC_API_KEY",
-    execution: "anthropic",
-    supportsDocuments: true,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-  },
-  zai: {
-    requiredEnv: "Z_AI_API_KEY",
-    execution: "openai-compatible",
-    supportsDocuments: false,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-    defaultBaseUrl: "https://api.z.ai/api/paas/v4",
-    forceChatCompletions: true,
-  },
-  nvidia: {
-    requiredEnv: "NVIDIA_API_KEY",
-    execution: "openai-compatible",
-    supportsDocuments: false,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-    defaultBaseUrl: "https://integrate.api.nvidia.com/v1",
-    forceChatCompletions: true,
-  },
-  minimax: {
-    requiredEnv: "MINIMAX_API_KEY",
-    execution: "openai-compatible",
-    supportsDocuments: false,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-    defaultBaseUrl: DEFAULT_MINIMAX_BASE_URL,
-    forceChatCompletions: true,
-  },
-  "github-copilot": {
-    requiredEnv: "GITHUB_TOKEN",
-    execution: "openai-http",
-    supportsDocuments: false,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-    defaultBaseUrl: GITHUB_MODELS_BASE_URL,
-    forceChatCompletions: true,
-  },
-  ollama: {
-    requiredEnv: "OLLAMA_BASE_URL",
-    execution: "openai-compatible",
-    supportsDocuments: false,
-    supportsStreaming: true,
-    supportsVideoUnderstanding: false,
-    defaultBaseUrl: DEFAULT_OLLAMA_BASE_URL,
-    forceChatCompletions: true,
-  },
-};
 
 export type ProviderRuntimeBindings = {
   apiKeys: Partial<Record<GatewayProvider, string | null>>;
@@ -180,112 +58,7 @@ export function resolveProviderOpenAiOverrides({
   };
 }
 
-type CliRequiredModelEnv = Extract<RequiredModelEnv, `CLI_${string}`>;
-
-export type CliProviderProfile = {
-  requiredEnv: CliRequiredModelEnv;
-  defaultModel: string | null;
-  missingBinaryLabel: string;
-  installLabel: string;
-  pathEnv: string;
-};
-
-const CLI_PROVIDER_PROFILES: Record<CliProvider, CliProviderProfile> = {
-  claude: {
-    requiredEnv: "CLI_CLAUDE",
-    defaultModel: "sonnet",
-    missingBinaryLabel: "Claude CLI",
-    installLabel: "Claude CLI",
-    pathEnv: "CLAUDE_PATH",
-  },
-  codex: {
-    requiredEnv: "CLI_CODEX",
-    defaultModel: null,
-    missingBinaryLabel: "Codex CLI",
-    installLabel: "Codex CLI",
-    pathEnv: "CODEX_PATH",
-  },
-  gemini: {
-    requiredEnv: "CLI_GEMINI",
-    defaultModel: "flash",
-    missingBinaryLabel: "Gemini CLI",
-    installLabel: "Gemini CLI",
-    pathEnv: "GEMINI_PATH",
-  },
-  agent: {
-    requiredEnv: "CLI_AGENT",
-    defaultModel: "auto",
-    missingBinaryLabel: "Cursor Agent CLI",
-    installLabel: "Cursor CLI",
-    pathEnv: "AGENT_PATH",
-  },
-  openclaw: {
-    requiredEnv: "CLI_OPENCLAW",
-    defaultModel: "main",
-    missingBinaryLabel: "OpenClaw CLI",
-    installLabel: "OpenClaw CLI",
-    pathEnv: "OPENCLAW_PATH",
-  },
-  opencode: {
-    requiredEnv: "CLI_OPENCODE",
-    defaultModel: null,
-    missingBinaryLabel: "OpenCode CLI",
-    installLabel: "OpenCode CLI",
-    pathEnv: "OPENCODE_PATH",
-  },
-  copilot: {
-    requiredEnv: "CLI_COPILOT",
-    defaultModel: null,
-    missingBinaryLabel: "GitHub Copilot CLI",
-    installLabel: "Copilot CLI",
-    pathEnv: "COPILOT_PATH",
-  },
-  agy: {
-    requiredEnv: "CLI_AGY",
-    defaultModel: null,
-    missingBinaryLabel: "Antigravity CLI",
-    installLabel: "agy",
-    pathEnv: "AGY_PATH",
-  },
-  pi: {
-    requiredEnv: "CLI_PI",
-    defaultModel: null,
-    missingBinaryLabel: "pi CLI",
-    installLabel: "pi",
-    pathEnv: "PI_PATH",
-  },
-};
-
-export const DEFAULT_CLI_MODELS = Object.fromEntries(
-  Object.entries(CLI_PROVIDER_PROFILES).map(([provider, profile]) => [
-    provider,
-    profile.defaultModel,
-  ]),
-) as Record<CliProvider, string | null>;
-
-export const DEFAULT_AUTO_CLI_ORDER: CliProvider[] = [
-  "claude",
-  "gemini",
-  "codex",
-  "agent",
-  "openclaw",
-  "opencode",
-  "copilot",
-  // agy is intentionally excluded from the default auto-fallback order.
-  // Use --cli agy or --model cli/agy to opt in explicitly.
-  // pi is also excluded; use --cli pi or --model cli/pi explicitly.
-];
-
-export function parseCliProviderName(raw: string): CliProvider | null {
-  const normalized = raw.trim().toLowerCase();
-  return Object.hasOwn(CLI_PROVIDER_PROFILES, normalized) ? (normalized as CliProvider) : null;
-}
-
-export function getCliProviderProfile(provider: CliProvider): CliProviderProfile {
-  return CLI_PROVIDER_PROFILES[provider];
-}
-
-export function requiredEnvForCliProvider(provider: CliProvider): RequiredModelEnv {
+export function requiredEnvForCliProvider(provider: CliProvider): CliRequiredModelEnv {
   return getCliProviderProfile(provider).requiredEnv;
 }
 
@@ -309,15 +82,7 @@ export function formatMissingCliModelError({
   return `${profile.missingBinaryLabel} not found for model ${userModelId}. Install ${profile.installLabel} or set ${profile.pathEnv}.`;
 }
 
-export function isGatewayProvider(provider: string): provider is GatewayProvider {
-  return Object.hasOwn(GATEWAY_PROVIDER_PROFILES, provider);
-}
-
-export function getGatewayProviderProfile(provider: GatewayProvider): GatewayProviderProfile {
-  return GATEWAY_PROVIDER_PROFILES[provider];
-}
-
-export function requiredEnvForGatewayProvider(provider: GatewayProvider): RequiredModelEnv {
+export function requiredEnvForGatewayProvider(provider: GatewayProvider): GatewayRequiredModelEnv {
   return getGatewayProviderProfile(provider).requiredEnv;
 }
 
@@ -422,52 +187,22 @@ export function resolveOpenAiCompatibleClientConfigForProvider({
       requestOptions,
     });
   }
-  if (provider === "github-copilot") {
-    const apiKey = openaiApiKey;
-    if (!apiKey) {
-      throw new Error("Missing GITHUB_TOKEN (or GH_TOKEN) for github-copilot/... model");
-    }
-    return {
-      apiKey,
-      baseURL: openaiBaseUrlOverride ?? GITHUB_MODELS_BASE_URL,
-      useChatCompletions: true,
-      isOpenRouter: false,
-      extraHeaders: buildGitHubModelsHeaders(),
-    };
-  }
-  if (provider === "ollama") {
-    return {
-      apiKey: openaiApiKey?.trim() || "ollama",
-      baseURL: openaiBaseUrlOverride ?? DEFAULT_OLLAMA_BASE_URL,
-      useChatCompletions: true,
-      isOpenRouter: false,
-      ...(requestOptions ? { requestOptions } : {}),
-    };
-  }
-
-  const apiKey = openaiApiKey;
+  const profile = GATEWAY_PROVIDER_PROFILES[provider];
+  const apiKey = provider === "ollama" ? openaiApiKey?.trim() || "ollama" : openaiApiKey;
   if (!apiKey) {
-    throw new Error(
-      provider === "zai"
-        ? "Missing Z_AI_API_KEY for zai/... model"
-        : provider === "minimax"
-          ? "Missing MINIMAX_API_KEY for minimax/... model"
-          : "Missing NVIDIA_API_KEY for nvidia/... model",
-    );
+    const requiredEnv =
+      provider === "github-copilot" ? "GITHUB_TOKEN (or GH_TOKEN)" : profile.requiredEnv;
+    throw new Error(`Missing ${requiredEnv} for ${provider}/... model`);
   }
-
-  const defaultBaseUrl =
-    provider === "zai"
-      ? "https://api.z.ai/api/paas/v4"
-      : provider === "minimax"
-        ? DEFAULT_MINIMAX_BASE_URL
-        : "https://integrate.api.nvidia.com/v1";
-
   return {
     apiKey,
-    baseURL: openaiBaseUrlOverride ?? defaultBaseUrl,
-    useChatCompletions: true,
+    baseURL: openaiBaseUrlOverride ?? profile.defaultBaseUrl,
+    useChatCompletions: profile.forceChatCompletions,
     isOpenRouter: false,
-    ...(requestOptions ? { requestOptions } : {}),
+    ...(provider === "github-copilot"
+      ? { extraHeaders: buildGitHubModelsHeaders() }
+      : requestOptions
+        ? { requestOptions }
+        : {}),
   };
 }
